backend/notification-service/requirements.txt
fastapi
uvicorn
twilio
python-dotenv
pydantic
